Consolidate etcd test client creation into pkg/etcdtest
Tests that need etcd had ~7 separate places creating their own
clients, with 3 different setupTestEtcd functions that were
basically the same code copy-pasted with slight variations. Worse,
when these tests ran outside the dev environment they'd hang
indefinitely because gRPC connections are lazy and the first real
RPC would block forever.

New pkg/etcdtest.TestEtcdClient(t) handles connection, fast-fail
with a clear error message, random prefix isolation, and cleanup
in one place. All existing etcd test setup functions now delegate
to it.
