+++ pkg/entity/testutils/inmem_server.go
@@ -7,14 +7,13 @@ import (
 	"strings"
 	"sync/atomic"
 	"testing"
-	"time"
 
 	clientv3 "go.etcd.io/etcd/client/v3"
 	apiserver "miren.dev/runtime/api/entityserver"
 	"miren.dev/runtime/api/entityserver/entityserver_v1alpha"
 	"miren.dev/runtime/pkg/entity"
 	"miren.dev/runtime/pkg/entity/schema"
-	"miren.dev/runtime/pkg/idgen"
+	"miren.dev/runtime/pkg/etcdtest"
 	"miren.dev/runtime/pkg/rpc"
 	"miren.dev/runtime/servers/entityserver"
 )
@@ -118,63 +117,36 @@ type EtcdEntityServer struct {
 	etcd   *clientv3.Client
 }
 
-// NewEtcdEntityServer creates a new etcd-backed entity server for testing
-// It connects to etcd:2379 and uses a random prefix for isolation
-// This should be run with ./hack/dev-exec or similar to have etcd available
+// NewEtcdEntityServer creates a new etcd-backed entity server for testing.
+// It connects to etcd:2379 and uses a random prefix for isolation.
+// Fails fast with a clear message if etcd is not reachable (i.e. outside dev env).
 func NewEtcdEntityServer(t *testing.T) (*EtcdEntityServer, func()) {
+	etcdClient, prefix := etcdtest.TestEtcdClient(t)
+
 	ctx := context.Background()
 	log := slog.New(slog.NewTextHandler(io.Discard, &slog.HandlerOptions{Level: slog.LevelError}))
 
-	// Connect to etcd
-	etcdClient, err := clientv3.New(clientv3.Config{
-		Endpoints:   []string{"etcd:2379"},
-		DialTimeout: 5 * time.Second,
-	})
-	if err != nil {
-		t.Fatalf("failed to connect to etcd: %v", err)
-	}
-
-	// Generate random prefix for isolation
-	prefix := "/" + idgen.Gen("test")
-
-	// Create etcd-backed store
 	store, err := entity.NewEtcdStore(ctx, log, etcdClient, prefix)
 	if err != nil {
-		etcdClient.Close()
 		t.Fatalf("failed to create etcd store: %v", err)
 	}
 
-	// Apply schema to the store
 	err = schema.Apply(ctx, store)
 	if err != nil {
-		etcdClient.Close()
 		t.Fatalf("failed to apply schema: %v", err)
 	}
 
-	// Create entity server
 	es, err := entityserver.NewEntityServer(log, store)
 	if err != nil {
-		etcdClient.Close()
 		t.Fatalf("failed to create entity server: %v", err)
 	}
 
-	// Create entity access client with local transport
 	localClient := rpc.LocalClient(entityserver_v1alpha.AdaptEntityAccess(es))
 	eac := entityserver_v1alpha.NewEntityAccessClient(localClient)
-
-	// Create the high-level entityserver client
 	client := apiserver.NewClient(log, eac)
 
-	cleanup := func() {
-		// Delete all keys with this prefix
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-		_, err := etcdClient.Delete(ctx, prefix, clientv3.WithPrefix())
-		if err != nil {
-			t.Logf("warning: failed to cleanup etcd prefix %s: %v", prefix, err)
-		}
-		etcdClient.Close()
-	}
+	// TestEtcdClient already registers prefix cleanup and client.Close
+	cleanup := func() {}
 
 	return &EtcdEntityServer{
 		Store:  store,
